finish task

+++ app/src/main/java/com/example/taskb/presentation/imageloaderdisk/ImageDownloader.kt
@@ -10,6 +10,8 @@ import java.io.FileOutputStream
 
 object ImageDownloader {
 
+    private const val IMAGES_FOLDER = "images"
+
     fun saveImagesToInternal(context: Context) {
         val imageIds = listOf(
             R.drawable.image21, R.drawable.image22, R.drawable.image23,
@@ -18,25 +20,35 @@ object ImageDownloader {
             R.raw.image11, R.raw.image12
         )
 
-        val imagesDir = File(context.filesDir, "images").apply { mkdirs() }
-        imageIds.forEachIndexed { index, resId ->
-            val drawable = ContextCompat.getDrawable(context, resId)
-            if (drawable is BitmapDrawable) {
-                val bitmap = drawable.bitmap
-                val file = File(imagesDir, "image_$index.jpg")
+        val imagesDir = File(context.filesDir, IMAGES_FOLDER).apply { mkdirs() }
 
-                FileOutputStream(file).use { out ->
-                    bitmap.compress(Bitmap.CompressFormat.JPEG, 100, out)
+        imageIds.forEachIndexed { index, resId ->
+            val file = File(imagesDir, "image_$index.jpg")
+            if (!file.exists()) {  // ✅ التحقق قبل الحفظ
+                val drawable = ContextCompat.getDrawable(context, resId)
+                if (drawable is BitmapDrawable) {
+                    val bitmap = drawable.bitmap
+                    FileOutputStream(file).use { out ->
+                        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, out)
+                    }
                 }
             }
         }
+
         svgIds.forEachIndexed { index, resId ->
             val file = File(imagesDir, "image_svg_$index.svg")
-            context.resources.openRawResource(resId).use { inputStream ->
-                file.outputStream().use { outputStream ->
-                    inputStream.copyTo(outputStream)
+            if (!file.exists()) {  // ✅ التحقق قبل الحفظ
+                context.resources.openRawResource(resId).use { inputStream ->
+                    file.outputStream().use { outputStream ->
+                        inputStream.copyTo(outputStream)
+                    }
                 }
             }
         }
+    }
+
+    fun isImagesSaved(context: Context): Boolean {
+        val imagesDir = File(context.filesDir, IMAGES_FOLDER)
+        return imagesDir.exists() && imagesDir.listFiles()?.isNotEmpty() == true
     }
 }
